+++ src/main/java/com/karnyshov/bsuirhub/model/service/GroupService.java
@@ -30,11 +30,11 @@ public interface GroupService {
      * @param criteria instance of {@link GroupFilterCriteria} that specifies a criteria for filtering.
      * @param keyword stored entities will be filtered by this value.
      * @param result {@link List} instance to hold retrieved entities.
-     * @return {@code long} value that represents count of ALL entities (even beyond specified bounds) satisfied
+     * @return {@code int} value that represents count of ALL entities (even beyond specified bounds) satisfied
      * filter criteria and filter value.
      * @throws ServiceException if an error occurred executing the method.
      */
-    long filter(int start, int size, GroupFilterCriteria criteria, String keyword, List<Group> result)
+    int filter(int start, int size, GroupFilterCriteria criteria, String keyword, List<Group> result)
             throws ServiceException;
 
     /**
@@ -43,10 +43,10 @@ public interface GroupService {
      * @param start lower bound index from which the result collection will start.
      * @param size size of the result collection.
      * @param result {@link List} instance to hold retrieved entities.
-     * @return {@code long} value that represents count of ALL entities (even beyond specified bounds).
+     * @return {@code int} value that represents count of ALL entities (even beyond specified bounds).
      * @throws ServiceException if an error occurred executing the method.
      */
-    long filter(int start, int size, List<Group> result) throws ServiceException;
+    int filter(int start, int size, List<Group> result) throws ServiceException;
 
     /**
      * Check if the name is unique.
